Move createMk3Slow() to SwerveModuleHelpers.java for better organization
