+++ src/main/java/frc/robot/utils/SwerveModuleHelpers.java
@@ -1,0 +1,69 @@
+package frc.robot.utils;
+
+import java.util.Optional;
+
+import org.northernforce.encoders.NFRCANCoder;
+import org.northernforce.motors.MotorEncoderMismatchException;
+import org.northernforce.motors.NFRTalonFX;
+import org.northernforce.subsystems.drive.swerve.NFRSwerveModule;
+import org.northernforce.subsystems.drive.swerve.NFRSwerveModule.Mk3SwerveConstants;
+import org.northernforce.subsystems.drive.swerve.NFRSwerveModule.NFRSwerveModuleConfiguration;
+
+import com.ctre.phoenix6.configs.TalonFXConfiguration;
+import com.ctre.phoenix6.signals.NeutralModeValue;
+
+import edu.wpi.first.math.system.plant.DCMotor;
+
+public class SwerveModuleHelpers
+{
+    /**
+     * Creates a traditional Mk3 swerve module with standard gearing and two falcon 500s, as well as a cancoder.
+     * @param name the name of the module
+     * @param driveID the id of the drive motor.
+     * @param turnID the id of the turn motor.
+     * @param cancoderID the id of the cancoder.
+     * @param invertDrive whether to invert the drive motor.
+     * @param canbus the name of the canbus
+     * @return a new NFRSwerveModule
+     */
+    public static NFRSwerveModule createMk3Slow(String name, int driveID, int turnID, int cancoderID, boolean invertDrive, String canbus)
+    {
+        NFRSwerveModuleConfiguration config = new NFRSwerveModuleConfiguration(name)
+            .withGearRatios(Mk3SwerveConstants.kDriveGearRatioSlow, Mk3SwerveConstants.kTurnGearRatio)
+            .withGearboxes(DCMotor.getFalcon500(1), DCMotor.getFalcon500(1))
+            .withMOIs(1.2, 1.2)
+            .withMaxSpeed(Mk3SwerveConstants.kDriveMaxSpeed);
+        TalonFXConfiguration driveConfig = new TalonFXConfiguration();
+        driveConfig.CurrentLimits.SupplyCurrentLimit = 60;
+        driveConfig.CurrentLimits.SupplyCurrentThreshold = 90;
+        driveConfig.CurrentLimits.SupplyTimeThreshold = 0.5;
+        driveConfig.CurrentLimits.SupplyCurrentLimitEnable = true;
+        driveConfig.Slot0.kP = Mk3SwerveConstants.kDriveP;
+        driveConfig.MotorOutput.NeutralMode = NeutralModeValue.Brake;
+        driveConfig.MotorOutput.DutyCycleNeutralDeadband = 0.1;
+        NFRTalonFX driveMotor = new NFRTalonFX(canbus, driveConfig, driveID);
+        driveMotor.getSelectedEncoder().setConversionFactor(Mk3SwerveConstants.kWheelCircumference /
+            Mk3SwerveConstants.kDriveGearRatioSlow);
+        driveMotor.setInverted(invertDrive);
+        TalonFXConfiguration turnConfig = new TalonFXConfiguration();
+        turnConfig.CurrentLimits.SupplyCurrentLimit = 60;
+        turnConfig.CurrentLimits.SupplyCurrentThreshold = 90;
+        turnConfig.CurrentLimits.SupplyTimeThreshold = 0.5;
+        turnConfig.CurrentLimits.SupplyCurrentLimitEnable = true;
+        turnConfig.Slot0.kP = Mk3SwerveConstants.kTurnP;
+        turnConfig.MotorOutput.NeutralMode = NeutralModeValue.Brake;
+        driveConfig.OpenLoopRamps.DutyCycleOpenLoopRampPeriod = 0.4;
+        turnConfig.ClosedLoopGeneral.ContinuousWrap = true;
+        NFRTalonFX turnMotor = new NFRTalonFX(canbus, turnConfig, turnID);
+        NFRCANCoder cancoder = new NFRCANCoder(canbus, cancoderID);
+        try
+        {
+            turnMotor.setSelectedEncoder(cancoder);
+        }
+        catch (MotorEncoderMismatchException e)
+        {
+            e.printStackTrace();
+        }
+        return new NFRSwerveModule(config, driveMotor, turnMotor, Optional.empty());
+    }
+}
